+++ src/main/java/club/iananderson/seasonhud/client/minimaps/JourneyMap.java
@@ -18,13 +18,9 @@ import net.minecraftforge.fml.ModList;
 import java.util.ArrayList;
 
 import static club.iananderson.seasonhud.SeasonHUD.MODID;
-import static club.iananderson.seasonhud.client.Calendar.calendar;
+import static club.iananderson.seasonhud.impl.sereneseasons.Calendar.calendar;
 import static club.iananderson.seasonhud.config.Config.enableMod;
-import static club.iananderson.seasonhud.data.CurrentSeason.*;
-
-/*Todo
-    * Need to switch names over to translatable ones
- */
+import static club.iananderson.seasonhud.impl.sereneseasons.CurrentSeason.*;
 
 public class JourneyMap {