+++ CHANGELOG.md
@@ -4,7 +4,17 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
+## [1.4.0] - 2023-01-31
+
+### Added
+
+- Added optional support for CuriosAPI
+    - If CuriosAPI is loaded, then the Calendar item from SereneSeason can be equipped in the "Charm" slot.
+    - If the "Need Calendar" option is enabled, then a Calendar in the "Charm" slot will meet the requirement
+
+
 ## [1.3.5] - 2023-01-22
+
 ### Changed
 - Synced version numbers back up with the main 1.19 branch
 - Slight reorganization to match the 1.19 version's file structure